requirements.txt
streamlit
streamlit-mic-recorder
SpeechRecognition
langchain
langchain-openai
langchain-google-genai
python-dotenv
pydub
